+++ changelog.md
@@ -1,3 +1,9 @@
+2019-07-22 - 0.9.91
+
+- resolve #47 about adding `--all-required` flag.
+  The new `--all-required` flag strictly forces all entries (even objects)
+  to be required (even the `@optional` annotation is ignored) 
+
 2019-02-01 - 0.9.9
 
 - publish v0.9.9 with new `--durations` flag to generate